feat: create a parallel route for Inaccuracy report modal
