+++ app/collection/items/[slug]/@modal/default.tsx
@@ -1,0 +1,3 @@
+export default function DefaultModalSlot() {
+  return null
+}
